Tidy up tests for extended model
