+++ tests/conftest.py
@@ -2,62 +2,14 @@ import pytest
 import pandas as pd
 import networkx as nx
 import numpy as np
-import numpy.typing as npt
 
 import torch
+import torch.nn as nn
 
 from grace.base import GraphAttrs, graph_from_dataframe
-from grace.io.core import Annotation
 from pathlib import Path
 
-
-def random_image_and_graph(
-    rng,
-    *,
-    num_nodes: int = 4,
-    image_size: tuple[int] = (128, 128),
-    feature_ndim: int = 32,
-) -> tuple[npt.NDArray, list[nx.Graph]]:
-    """Create a random image and graph."""
-    image = np.zeros(image_size, dtype=np.uint16)
-
-    features = [rng.uniform(size=(feature_ndim,)) for _ in range(num_nodes)]
-
-    node_coords = rng.integers(0, image.shape[1], size=(num_nodes, 2))
-    node_ground_truth = rng.choice(
-        [Annotation.TRUE_NEGATIVE, Annotation.TRUE_POSITIVE], size=(num_nodes,)
-    )
-    df = pd.DataFrame(
-        {
-            GraphAttrs.NODE_X: node_coords[:, 0],
-            GraphAttrs.NODE_Y: node_coords[:, 1],
-            GraphAttrs.NODE_FEATURES: features,
-            GraphAttrs.NODE_GROUND_TRUTH: node_ground_truth,
-            GraphAttrs.NODE_CONFIDENCE: rng.uniform(
-                size=(num_nodes),
-            ),
-        }
-    )
-
-    image[tuple(node_coords[:, 1]), tuple(node_coords[:, 0])] = 1
-    graph = graph_from_dataframe(df, triangulate=True)
-
-    graph.update(
-        edges=[
-            (
-                src,
-                dst,
-                {
-                    GraphAttrs.EDGE_GROUND_TRUTH: rng.choice(
-                        [Annotation.TRUE_NEGATIVE, Annotation.TRUE_POSITIVE],
-                    )
-                },
-            )
-            for src, dst in graph.edges
-        ]
-    )
-
-    return image, graph
+from _utils import random_image_and_graph
 
 
 @pytest.fixture(scope="session")
@@ -120,11 +72,11 @@ def mrc_image_and_annotations_dir(tmp_path_factory, default_rng) -> Path:
     return tmp_data_dir
 
 
-class SimpleExtractor(torch.nn.Module):
+class SimpleExtractor(nn.Module):
     def forward(self, x):
         return torch.rand(x.size(0), 2)
 
 
 @pytest.fixture(scope="session")
-def simple_extractor() -> torch.nn.Module:
+def simple_extractor() -> nn.Module:
     return SimpleExtractor()
